megano/pyproject.toml
[tool.poetry]
name = "megano"
version = "0.1.0"
description = ""
authors = ["ADMIN <k.safelkin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
django = "4.1.5"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
django-filter = "^23.5"
django-mptt = "^0.15.0"
django-request-logging = "^0.7.5"
gunicorn = "^21.2.0"
pillow = "^10.2.0"
drf-spectacular = "^0.27.0"
typing-extensions = "^4.9.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
